+++ CHANGELOG.md
@@ -76,6 +76,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.35] - 2026-03-30
+
+### Docs
+- Update README.md
+- Update TODO.md
+- Update docs/README.md
+
+### Test
+- Update tests/test_llx_mcp.py
+
+### Other
+- Update pyqual/integrations/llx_mcp.py
+
 ## [0.1.34] - 2026-03-30
 
 ### Test